Optimize pipeline to run certain stages for relevant changes (#59)
Pipeline only runs certain stages when the relevant directories are
changed.

Additional features:
- Added a build check for the backend
- Added feature for being able to skip all runs or force all runs by
adding `[skip]` or `[run]` to your commit message. Can be anywhere in
the commit message but preferably at the end, so the main commit message
is more readable for others.
  
This also only supports the first commit message given to git. If you
try to add to the message body by doing something like:

  ```sh
  git commit -m "this is a commit" -m "this is in the body [run]"
  ```

  The `[run]` will not be read. Only `this is a commit` will be checked.

Closes #56